Cloud_1/requirements.txt
selenium==4.21.0
google-cloud-firestore==2.21.0      # fornece o pacote Python "google"
pydantic==2.7.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
structlog==24.1.0
fastapi~=0.115.14
schedule~=1.2.2